fix(tours): total_eqp fractionnaire (fix "Echec sauvegarde brouillon")
La sauvegarde d'un brouillon de tour echouait (422 affiche "[object Object]")
quand les volumes injectes donnaient un EQP total fractionnaire : total_eqp
etait type int (schema + colonne) alors que tour_stops.eqp_count est deja
numeric(10,2). Pydantic v2 rejette un float a partie fractionnaire sur un int.

- schema TourBase/TourUpdate.total_eqp : int -> float
- modele Tour.total_eqp : Integer -> Numeric(10,2)
- _migrate_column_types : conversion auto tours.total_eqp -> numeric(10,2)
- util getApiErrorMessage : extrait un message lisible des erreurs 422
  (detail tableau) au lieu de "[object Object]" ; applique a TourBuilder
- tests/test_tours.py : garde-fou total_eqp fractionnaire

+++ Chaos_Route/backend/app/models/tour.py
@@ -37,7 +37,7 @@ class Tour(Base):
     return_time: Mapped[str | None] = mapped_column(String(5))  # HH:MM
     total_km: Mapped[float | None] = mapped_column(Numeric(10, 2))
     total_duration_minutes: Mapped[int | None] = mapped_column(Integer)
-    total_eqp: Mapped[int | None] = mapped_column(Integer)
+    total_eqp: Mapped[float | None] = mapped_column(Numeric(10, 2))  # EQP fractionnaire (somme eqp_count)
     total_cost: Mapped[float | None] = mapped_column(Numeric(12, 2))
     total_weight_kg: Mapped[float | None] = mapped_column(Numeric(10, 2))
     # Poids total du tour (saisi par le postier) / Total tour weight (entered by dispatcher)
